Consolidated test nuget package and reordering of folders to make easier to find tests and common classes.
