+++ AcceptanceTests/AcceptanceTests.Tests/IntegrationTests/DriverHookTests.cs
@@ -2,7 +2,8 @@
 using AcceptanceTests.Driver;
 using AcceptanceTests.Driver.Settings;
 using AcceptanceTests.Model.Context;
-using AcceptanceTests.Tests.Hooks;
+using AcceptanceTests.SpecflowTests.Common;
+using AcceptanceTests.SpecflowTests.Common.Hooks;
 using Coypu;
 using FluentAssertions;
 using NUnit.Framework;
