+++ AcceptanceTests/AcceptanceTests.Tests/IntegrationTests/HookTestsBase.cs
@@ -1,5 +1,5 @@
-﻿using BoDi;
-using NUnit.Framework;
+﻿using AcceptanceTests.SpecflowTests.Common;
+using BoDi;
 
 namespace AcceptanceTests.Tests.IntegrationTests
 {
