+++ AcceptanceTests/AcceptanceTests.Tests/IntegrationTests/NUnitParamReaderTests.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using AcceptanceTests.Driver.Support;
 using AcceptanceTests.Model;
+using AcceptanceTests.SpecflowTests.Common;
 using FluentAssertions;
 using NUnit.Framework;
