+++ AcceptanceTests/AcceptanceTests.Tests/SpecflowTests/AdminWebsite/Resources/adminwebsite.safari.desktop.runsettings.xml
