+++ AcceptanceTests/AcceptanceTests.Tests/SpecflowTests/Common/Steps/ErrorSteps.cs
@@ -7,7 +7,7 @@ using Coypu;
 using FluentAssertions;
 using TechTalk.SpecFlow;
 
-namespace AcceptanceTests.Tests.Steps
+namespace AcceptanceTests.SpecflowTests.Common.Steps
 {
     [Binding]
     public class ErrorSteps : StepsBase
