+++ AcceptanceTests/AcceptanceTests.Tests/SpecflowTests/Common/Steps/QuestionnaireSteps.cs
@@ -2,7 +2,7 @@
 using Coypu;
 using TechTalk.SpecFlow;
 
-namespace AcceptanceTests.Tests.Steps
+namespace AcceptanceTests.SpecflowTests.Common.Steps
 {
     [Binding]
     public class QuestionnaireSteps : StepsBase
