+++ AcceptanceTests/AcceptanceTests.Tests/SpecflowTests/Common/Steps/StepsBase.cs
@@ -2,7 +2,7 @@
 using Coypu;
 using TechTalk.SpecFlow;
 
-namespace AcceptanceTests.Tests.Steps
+namespace AcceptanceTests.SpecflowTests.Common.Steps
 {
     public class StepsBase
     {
