--- AcceptanceTests/AcceptanceTests.Tests/resources/videowebsite/videowebsite.appsettings.json
@@ -1,3 +1,0 @@
-﻿{
-  "WebsiteUrl": "https://vh-video-web-sandbox.hearings.reform.hmcts.net/"
-}
